Extract ResidentialAddressPlugin module
Gonna share some code here.
